Fixing packaging + deleting unused classes
