+++ msal/src/main/java/com/microsoft/identity/client/IMultiTenantAccount.java
@@ -20,7 +20,7 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.client.tenantprofile;
+package com.microsoft.identity.client;
 
 import android.support.annotation.NonNull;
